render: xcursor_update(): convert local ‘scale’ variable to float

+++ render.c
@@ -4276,7 +4276,7 @@ render_xcursor_update(struct seat *seat)
 
     xassert(seat->pointer.cursor != NULL);
 
-    const int scale = seat->pointer.scale;
+    const float scale = seat->pointer.scale;
     struct wl_cursor_image *image = seat->pointer.cursor->images[0];
 
     wl_surface_attach(
